% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_scripts.R
\name{detect_scripts}
\alias{detect_scripts}
\title{Detect scripts in a folder}
\usage{
detect_scripts(base_path, exclude_status = c("LOCK"), recursive = TRUE)
}
\description{
Detect scripts in a folder
}
