% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_log.R
\name{extract_log}
\alias{extract_log}
\title{Extract log from script}
\usage{
extract_log(script_path, type = c("message", "output"))
}
\arguments{
\item{script_path}{Path to the script's directory}
}
\value{
Character vector of log messages
}
\description{
Extract log from script
}
