% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variable.R
\name{extract_variable}
\alias{extract_variable}
\title{Extract variable from a completed script}
\usage{
extract_variable(script_path, variable_name, drop = length(variable_name) == 1)
}
\arguments{
\item{script_path}{The path to the script}

\item{variable_name}{The name of the variable(s) to extract, character vector}
}
\value{
A named list with names equal to \code{variable_name}, or if \code{drop == TRUE}
and variable_name is of length 1, the value of the variable.
}
\description{
This function assumes that the script has completed execution without error.
}
