% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_script.R
\name{make_script}
\alias{make_script}
\title{Make a script that runs an expression}
\usage{
make_script(expr, name, path, seed, ...)
}
\arguments{
\item{expr}{The expression to evaluate in the script. Relevant global
and local variables, and packages are detected with \code{\link[future]{getGlobalsAndPackages}}.
Note that this means that \code{expr} should be able to run in the current session.}

\item{...}{Other arguments passed to \code{\link[future]{getGlobalsAndPackages}}}
}
\value{
The path to the folder that represents the script
}
\description{
A folder is created that contains the script code, and the variable environment
in which it is executed. This is deduced from the expression (\code{expr}) passed
to this function.
}
\examples{
path_to_script <- make_script({  }, name = 'test-1', path = 'test-scripts')

run_script(path_to_script)

}
