% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_script.R
\name{remove_script}
\alias{remove_script}
\title{Remove a script (directory)s}
\usage{
remove_script(script_path, askConfirmation = interactive())
}
\arguments{
\item{script_path}{The script's directory}

\item{askConfirmation}{Whether to ask for confirmation}
}
\description{
Remove a script (directory)s
}
