% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{run_script}
\alias{run_script}
\title{Run script made with \code{\link{make_script}}}
\usage{
run_script(
  script_path,
  rscript_binary = file.path(R.home("bin"), "Rscript"),
  wait = TRUE,
  ignore.stdout = TRUE,
  ignore.stderr = TRUE,
  show.output.on.console = FALSE,
  timeout = 0,
  ...
)
}
\arguments{
\item{script_path}{Path to the folder of the script}

\item{rscript_binary}{Location of rscript binary}

\item{wait}{see ?system}

\item{ignore.stdout}{see ?system}

\item{ignore.stderr}{see ?system}

\item{show.output.on.console}{see ?system}

\item{timeout}{see ?system}
}
\description{
Run script made with \code{\link{make_script}}
}
\examples{
run_script(make_script({ ... }, name = 'test-1', folder = 'test-scripts'))

}
