% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{run_script}
\alias{run_script}
\title{Run script made with \code{\link{make_script}}}
\usage{
run_script(script_path, wait = TRUE)
}
\arguments{
\item{script_path}{Path to the folder of the script}

\item{wait}{Whether to wait for the script to complete (Default) or not}
}
\description{
Run script made with \code{\link{make_script}}
}
\examples{
run_script(make_script({ ... }, name = 'test-1', folder = 'test-scripts'))

}
